refactor(studio): extract the schema explorer into a hook
SchemaViewer was 489 lines — 471 of logic behind 18 of markup — now 231.

`useSchemaExplorer` owns both schema planes: the shard's tables with their lazily
probed columns and indexes, the `.global()` (D1) tables with theirs, the graph's
column probe, and the view / shard / filter selection. The two planes live in one
hook on purpose: the graph probe needs both table lists, so separating them would
only mean threading one into the other.

Its return type is named rather than inferred, so the hook's surface is a
readable contract instead of twenty-odd inferred fields.

Suppression deleted. 930 studio tests green.

Refs #230.
